refactor example modules
