+++ example/bwmf/full_topo.go
@@ -1,4 +1,4 @@
-package example
+package bwmf
 
 //The full structure is basically assume that every one is also parent for every else.
 //And everyone else is communicating to get the data they need.
