+++ example/regression/tree_topo_test.go
@@ -1,4 +1,4 @@
-package example
+package regression
 
 import "testing"
