+++ integration/node_failure_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/coreos/go-etcd/etcd"
 	"github.com/taskgraph/taskgraph/controller"
-	"github.com/taskgraph/taskgraph/framework"
+	"github.com/taskgraph/taskgraph/example/regression"
 )
 
 // TestMasterSetEpochFailure checks if a master task failed at SetEpoch,
@@ -23,8 +23,7 @@ func TestMasterSetEpochFailure(t *testing.T) {
 	controller := controller.New(job, etcd.NewClient(etcdURLs), numOfTasks)
 	controller.Start()
 
-	// We need to set etcd so that nodes know what to do.
-	taskBuilder := &framework.SimpleTaskBuilder{
+	taskBuilder := &regression.SimpleTaskBuilder{
 		GDataChan:    make(chan int32, 10),
 		NodeProducer: make(chan bool, 1),
 		MasterConfig: map[string]string{
@@ -89,8 +88,7 @@ func testSlaveFailure(t *testing.T, job string, slaveConfig map[string]string) {
 	controller.Start()
 	defer controller.Stop()
 
-	// We need to set etcd so that nodes know what to do.
-	taskBuilder := &framework.SimpleTaskBuilder{
+	taskBuilder := &regression.SimpleTaskBuilder{
 		GDataChan:          make(chan int32, 10),
 		NodeProducer:       make(chan bool, 1),
 		SlaveConfig:        slaveConfig,
